+++ lib/openpanel_flutter.dart
@@ -3,4 +3,4 @@ library openpanel_flutter;
 export 'src/open_panel.dart';
 export 'src/models/open_panel_event_options.dart';
 export 'src/models/open_panel_options.dart';
-export 'src/route_observer.dart';
+export 'src/observers/route_observer.dart';