+++ lib/src/models/open_panel_event_options.dart
@@ -5,16 +5,6 @@ class OpenpanelEventOptions extends Equatable {
 
   const OpenpanelEventOptions({this.profileId});
 
-  factory OpenpanelEventOptions.fromJson(Map<String, dynamic> json) {
-    return OpenpanelEventOptions(profileId: json['profileId']);
-  }
-
   @override
   List<Object?> get props => [profileId];
-
-  Map<String, dynamic> toJson() {
-    return {
-      'profileId': profileId,
-    };
-  }
 }